Working FFT Code
Ujjwal, the file multiplier2.v was to only test the complex number multiplication. The complete FFT code is present in the file fft.v. The code works now with decent accuracy. However, the control signals need to implemented and simulation with GTK wave and power analysis need to be done.

And I have also updated the pdf that I used for all my calculations.